VANET_Misbehavior_XAI/VANET_Misbehavior_XAI/src/data_preprocessing/data_augmentation.py
# src/data_preprocessing/data_augmentation.py
import numpy as np
import pandas as pd
from sklearn.utils import resample

class VANETDataAugmentation:
    def __init__(self, random_state=42):
        self.random_state = random_state
    
    def augment_minority_class(self, X, y, target_class, ratio=1.0):
        """Augment minority class samples to balance the dataset"""
        # Implementation of class balancing
        return X, y
    
    def add_gaussian_noise(self, X, noise_factor=0.05):
        """Add Gaussian noise to features for robustness"""
        # Implementation of noise addition
        return X
    
    def generate_synthetic_attacks(self, X, attack_type):
        """Generate synthetic attack patterns based on existing data"""
        # Implementation of synthetic attack generation
        return X
