VANET_Misbehavior_XAI/VANET_Misbehavior_XAI/requirements.txt
tensorflow==2.20.0
torch==2.8.0
torchvision==0.23.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
shap==0.42.1
lime==0.2.0.1
plotly==5.15.0
dash==2.11.1
pyyaml==6.0.1
jupyter==1.0.0
tqdm==4.65.0
